boilerplate/pyproject.toml
[tool.poetry]
name = "boilerplate"
version = "0.1.0"
description = "Stock API using FastAPI"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{include = "boilerplate", from = "src"}]

[tool.poetry.dependencies]
python = "~3.12.1"
fastapi = "^0.109.1"
uvicorn = "^0.30.3"
python-dotenv = "^0.21.0"  
requests = "^2.26.0"  
pydantic = "^2.10.6"  
pytest = "^7.2.2" 

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
